+++ assets/scripts/domElements.js
@@ -2,6 +2,7 @@ const domElements = {
   inputSecondary: document.querySelectorAll('.input-secondary'),
   inputSecondaryIcon: document.querySelectorAll('.input-secondary-icon'),
   inputSecondaryAllItems: document.querySelectorAll('.input-secondary-items'),
+  allRecipes: document.querySelector('.recipes'),
 };
 
 export default domElements;
