+++ assets/scripts/pages/index.js
@@ -1,6 +1,14 @@
 import domElements from '../domElements.js';
-import getInputSecondaryElements from '../factory/inputSecondary.js';
+import clickSecondaryElementsArrow from '../actions/inputSecondary.js';
+import { showInputsSecondary, showRecipes } from '../data/showData.js';
 
 domElements.inputSecondaryIcon.forEach((icon) =>
-  icon.addEventListener('click', (e) => getInputSecondaryElements(e))
+  icon.addEventListener('click', (e) => clickSecondaryElementsArrow(e))
 );
+
+const init = () => {
+  showInputsSecondary();
+  showRecipes();
+};
+
+init();
